optimize dashboard components
