+++ src/Pages/Dashboard/components/TwinklingStars.tsx
@@ -1,0 +1,38 @@
+// components/dashboard/TwinklingStars.tsx
+import React from 'react';
+
+interface TwinklingStarsProps {
+  count?: number;
+}
+
+export const TwinklingStars: React.FC<TwinklingStarsProps> = ({ count = 20 }) => {
+  const stars = Array.from({ length: count }, (_, i) => ({
+    id: i,
+    size: Math.random() * 2 + 0.5,
+    x: Math.random() * 100,
+    y: Math.random() * 100,
+    delay: Math.random() * 4,
+    duration: Math.random() * 3 + 2,
+  }));
+
+  return (
+    <div className="absolute inset-0 pointer-events-none overflow-hidden">
+      {stars.map((star) => (
+        <div
+          key={star.id}
+          className="absolute rounded-full bg-white animate-pulse"
+          style={{
+            width: `${star.size}px`,
+            height: `${star.size}px`,
+            left: `${star.x}%`,
+            top: `${star.y}%`,
+            animationDelay: `${star.delay}s`,
+            animationDuration: `${star.duration}s`,
+          }}
+        />
+      ))}
+    </div>
+  );
+};
+
+export default TwinklingStars;
