+++ src/Pages/Dashboard/components/index.ts
@@ -1,0 +1,10 @@
+// Pages/Dashboard/Dashboard/index.ts
+export { TwinklingStars } from './TwinklingStars';
+export { DashboardHeader } from './DashboardHeader';
+export { WelcomeSection } from './WelcomeSection';
+export { UserInfoCards } from './UserInfoCards';
+export { StatCard } from './StatCard';
+export { StatisticsSection } from './StatisticsSection';
+export { ActionButton } from './ActionButton';
+export { ActionsSection } from './ActionsSection';
+export { DashboardFooter } from './DashboardFooter';
